requirements.txt
# === JODOHKU.MY Backend Dependencies ===
# FastAPI Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.2
greenlet==3.0.3

# Redis & Caching
redis==5.1.0
aioredis==2.0.1

# Validation & Serialization
pydantic==2.9.0
pydantic-settings==2.5.0
email-validator==2.2.0

# HTTP & External APIs
httpx==0.27.0
aiohttp==3.10.0

# Image Processing
Pillow==10.4.0
python-magic==0.4.27

# WebSocket
websockets==12.0

# Payment Gateway
toyyibpay-python==1.0.0

# Email
aiosmtplib==3.0.1
jinja2==3.1.4

# Task Queue
celery==5.4.0
flower==2.0.1

# Monitoring
sentry-sdk[fastapi]==2.14.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.27.0

# Utilities
python-dotenv==1.0.1
pytz==2024.1
uuid6==2024.7.10
nanoid==2.0.0
